address-options: move component to containers page
